refactor installations script. Extract utils, modify branches to point to master, add confirmation message and stash changes on init
